1. Code review: moved the navigation method to "ReminderScreenMethods.kt".
